requirements.txt
openai-whisper
torch==2.5.1+cu121
torchvision==0.20.1+cu121
--extra-index-url https://download.pytorch.org/whl/cu121
